+++ packages/core/src/utils/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @brayford/core/utils
+ * 
+ * Utility functions for Project Brayford
+ */
+
+export * from './email-domains';